use parent_class draw method for non printer painters
2001-02-09  Radek Doulik  <rodo@ximian.com>

	* htmliframe.c (draw): use parent_class draw method for non
	printer painters

	* htmlembedded.c: reworked size calculation/allocation, removed
	ugly hacks ;-)

	* gtkhtml-embedded.c (gtk_html_embedded_new): don't set
	resize_mode to QUEUE

	* gtkhtml.c (gtk_html_set_default_content_type): take care of
	content_type == NULL

svn path=/trunk/; revision=2821

+++ src/htmlembedded.h
@@ -33,7 +33,7 @@ struct _HTMLEmbedded {
 	gchar *value;
 	HTMLForm *form;
 	GtkWidget *widget, *parent;
-	gint16 width, height;
+	gint width, height;
 
 	gint abs_x, abs_y;
 };
